requirements.txt
jsonref
jsonschema
json-schema-for-humans
paho-mqtt
cbor
numpy
jinja2